[SPARK-20013][SQL]add a newTablePath parameter for renameTable in ExternalCatalog
